api/requirements.txt
# -- API --
fastapi
uvicorn

# -- Tratamento de arquivos --
unstructured[all-docs,image]
pytesseract

# -- Integração com IA --
#openai
#transformers
google-generativeai
werkzeug